Convert real to accreal in libTHCUNN

+++ SparseLinear.cu
@@ -34,8 +34,8 @@ void THNN_CudaHalfSparseLinear_accGradParameters(
           THCudaHalfTensor *gradBias,
           THCudaHalfTensor *weight,
           THCudaHalfTensor *bias,
-          double weightDecay,
-          double scale) {
+          float weightDecay,
+          float scale) {
   THError("THCudaHalfTensor not supported with SparseLinear");
 }
 
@@ -56,8 +56,8 @@ void THNN_CudaHalfSparseLinear_legacyAccGradParameters(
           THCudaHalfTensor *gradBias,
           THCudaHalfTensor *weight,
           THCudaHalfTensor *bias,
-          double weightDecay,
-          double scale) {
+          float weightDecay,
+          float scale) {
   THError("THCudaHalfTensor not supported with SparseLinear");
 }
 
@@ -76,7 +76,7 @@ void THNN_CudaHalfSparseLinear_updateParameters(
           THCudaHalfTensor *gradWeight,
           THCudaHalfTensor *gradBias,
           THCudaHalfTensor *lastInput,
-          double learningRate) {
+          float learningRate) {
   THError("THCudaHalfTensor not supported with SparseLinear");
 }
 #endif
